Update messages state explicitly
Since `response.data` is really an object with a messages attribute,
what you had totally works. I tend to prefer setting the state more
explicitly, especially in the future if your state becomes larger than
just the messages attribute. Or if say the API changes, it may be easier
to see what is going on.

If you switch to the ES6 syntax in the future, you could even do something like this:

const { messages } = response.data;
this.setState({ messages });

Since we use destructuring here, and the simplified object syntax in
`setState`, this actually is the same thing. But, I digress.