Move Message style logic to function
I like to keep logic outside of the render function if possible. If
there is no possibility of this value changing, you could actually
define this as a variable outside of the render function so that the
logic isn't run every time the component is re-rendered. However, I'll
leave it like this so you can demo some vulnerabilities ;)