Move form styles to class variables
Here's an example of just using variable to define style. You could do
an inline if statement here like:

inputStyle = {
  float: isTrue ? "right" : "left"
}

And now this will only get evaluated the first time the component is
instantiated.

Also, I like how this tends to clean up the JSX. I tend to move what I
can out of the return statement so it is only focused on the markup
representation.